undo footer
